requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.68.1
google-api-core==2.0.1
google-auth==2.2.0
google-cloud-core==2.0.0
google-cloud-datastore==2.1.6
googleapis-common-protos==1.53.0
grpcio==1.41.0
h11==0.12.0
idna==3.2
libcst==0.3.21
mypy-extensions==0.4.3
passlib==1.7.4
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
uvicorn==0.15.0
gunicorn==20.0.4
uvloop==0.15.2   
httptools==0.1.1